requirements.txt
streamlit==1.35.0
tensorflow-cpu==2.15.0
opencv-python-headless==4.8.0.76
Pillow==10.2.0
h5py==3.10.0
numpy==1.26.4
